Refactor FuneralFeatureTest using a trait
