+++ tests/Feature/FuneralFeatureTest.php
@@ -2,37 +2,15 @@
 
 namespace Tests\Feature;
 
-use App\City;
 use App\Funeral;
-use App\Http\Middleware\Authenticate;
-use App\User;
-use Spatie\Permission\Models\Permission;
+use App\Traits\TestWithCredentialsTrait;
 use Tests\TestCase;
-use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 
 class FuneralFeatureTest extends TestCase
 {
     use RefreshDatabase;
-
-    /** @var City */
-    protected $city;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-        $this->withoutMiddleware(Authenticate::class);
-        $this->app->make(\Spatie\Permission\PermissionRegistrar::class)->registerPermissions();
-
-        Permission::create(['name' => 'gd-bearbeiten']);
-
-        $this->city = factory(City::class)->create();
-
-        /** @var User */
-        $this->user = factory(User::class)->create();
-        $this->user->givePermissionTo('gd-bearbeiten');
-        $this->user->writableCities()->attach($this->city);
-    }
+    use TestWithCredentialsTrait;
 
     /**
      * Test that a funeral can be created
